Resolved Problem with Finding Vector
Changed CVector to a java file to allow the program to find it. Also
changed the other files back to pde files to allow them to use
processing’s functions.

+++ Project/Celestial.pde
@@ -4,7 +4,7 @@ class Celestial {
   CVector location; 
   CVector velocity; 
   CVector acceleration;
-  float gravConst = 6.67 * pow(10, -11);
+  float gravConst = (float) (6.67 * pow(10.0, -11.0));
 
   CVector netForce;
